+++ src/infrastructure/index.ts
@@ -1,0 +1,6 @@
+// Repositories
+export {
+  StaticExperienceRepository,
+  StaticProjectRepository,
+  StaticSkillRepository,
+} from './repositories';